requirements.txt
election-service-grpc @ git+https://github.com/CapyDevelop/grpc_election.git@main
greenlet==3.0.2
grpcio==1.60.0
grpcio-tools==1.60.0
orm-models @ git+https://github.com/alexseipopov/capybaras_orm_models.git@main
pika==1.3.2
protobuf==4.25.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
